politicas para career

+++ SistemaPracticas/app/Http/Controllers/CareerController.php
@@ -12,20 +12,28 @@ class CareerController extends Controller
 {
 
     public function index(){
+        $this->authorize('viewAny', Career::class);
         return new CareerCollection(Career::paginate());
     }
     public function show(Career $Career){
+        $this->authorize('view', $Career);
         return response()->json(new CareerResource($Career),200);
     }
     public function store(Request $request){
+        $this->authorize('create', Career::class);
+
         $Career= Career::create($request->all());
         return response()->json($Career,201);
     }
     public function update(Request $request,Career $Career){
+        $this->authorize('update', $Career);
+
         $Career->update($request->all());
         return response()->json($Career,200);
     }
     public function delete(Career $Career){
+        $this->authorize('delete', $Career);
+
         $Career->delete();
         return response()->json(null,204);
     }
